bookmarks

+++ app/Http/Controllers/JobController.php
@@ -5,16 +5,19 @@ namespace App\Http\Controllers;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
-use App\Models\Job;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Job;
 
 class JobController extends Controller
 {
+    use AuthorizesRequests;
+
     // @desc Show all job listing
     // @route GET /jobs
     public function index(): View
     {
-        $jobs = Job::all();
+        $jobs = Job::paginate(9);
         return view('jobs.index')->with('jobs', $jobs);
     }
 
@@ -79,6 +82,9 @@ class JobController extends Controller
     // @route GET /jobs/{$id}/edit
     public function edit(Job $job): View
     {
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         return view('jobs.edit')->with('job', $job);
     }
 
@@ -86,6 +92,9 @@ class JobController extends Controller
     // @route PUT /jobs/{$id}
     public function update(Request $request, Job $job): string
     {
+        // Check if user is authorized
+        $this->authorize('update', $job);
+
         $validatedData = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
@@ -128,14 +137,20 @@ class JobController extends Controller
     // @route DELETE /jobs/{$id}
     public function destroy(Job $job): RedirectResponse
     {
-        // If logo, then delete it
+        // Check if user is authorized
+        $this->authorize('delete', $job);
 
+        // If logo, then delete it
         if ($job->company_logo) {
             Storage::delete('public/logos/' . basename($job->company_logo));
         }
 
         $job->delete();
 
+        // Check if request came from the dashboard
+        if (request()->query('from') == 'dashboard') {
+            return redirect()->route('dashboard')->with('success', 'Job listing deleted successfully!');
+        }
         return redirect()->route('jobs.index')->with('success', 'Job listing deleted successfully!');
     }
 }
